requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
elevenlabs==1.50.7
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.18
langchain-core==0.3.34
langchain-ollama==0.2.3
langchain-text-splitters==0.3.6
langgraph==0.2.70
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.3.7
msgpack==1.1.0
multidict==6.1.0
numpy==1.26.4
ollama==0.4.7
orjson==3.10.15
packaging==24.2
propcache==0.2.1
PyAudio==0.2.14
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SpeechRecognition==3.14.1
SQLAlchemy==2.0.38
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.2
yarl==1.18.3
zstandard==0.23.0
